test-td-payload: refine the test crate
Simplify implementation of the test-td-payload crate.
