+++ tests/test-td-paging/src/lib.rs
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
 #![no_std]
+#![cfg_attr(test, no_main)]
 // The `custom_test_frameworks` feature allows the use of `#[test_case]` and `#![test_runner]`.
 // Any function, const, or static can be annotated with `#[test_case]` causing it to be aggregated
 // (like #[test]) and be passed to the test runner determined by the `#![test_runner]` crate
@@ -11,17 +12,16 @@
 #![test_runner(test_runner)]
 // Reexport the test harness main function under a different symbol.
 #![reexport_test_harness_main = "test_main"]
-#![cfg_attr(test, no_main)]
 
 #[cfg(test)]
-entry_point!(kernel_main);
-
-#[cfg(test)]
-use bootloader::{boot_info as info, entry_point, BootInfo};
+use bootloader::{boot_info, entry_point, BootInfo};
 #[cfg(test)]
 use core::ops::Deref;
 #[cfg(test)]
 use test_runner_client::{init_heap, serial_println, test_runner};
+
+#[cfg(test)]
+entry_point!(kernel_main);
 
 #[cfg(test)]
 fn kernel_main(boot_info: &'static mut BootInfo) -> ! {
@@ -36,8 +36,8 @@ fn kernel_main(boot_info: &'static mut BootInfo) -> ! {
     let offset = boot_info.physical_memory_offset.into_option().unwrap();
 
     for usable in memoryregions.iter() {
-        if usable.kind == info::MemoryRegionKind::Usable {
-            init_heap((usable.start + offset) as usize, 0x10000);
+        if usable.kind == boot_info::MemoryRegionKind::Usable {
+            init_heap((usable.start + offset) as usize, 0x100000);
             break;
         }
     }
