--- tests/test-td-payload/.gitignore
@@ -1,7 +1,0 @@
-clear-*-kvm.img*
-fat*.img
-target
-test_data
-Cargo.lock
-*.o
-*.obj
